LayoutTests/platform/glib/accessibility/aria-roledescription-expected.txt

This tests that aria-roledescription works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axButton.roleDescription is 'AXRoleDescription: Super Button'
PASS Got default role description: AXRoleDescription:
PASS axButton.roleDescription === 'AXRoleDescription: Super Button'
PASS Got default role description: AXRoleDescription:
PASS axButton.roleDescription === 'AXRoleDescription: Super Button'
PASS Got default role description: AXRoleDescription:
PASS successfullyParsed is true

TEST COMPLETE
text
